+++ functions/src/functions/v1/slack/index.ts
@@ -1,4 +1,5 @@
 export { clearVerificationCode as clear_verification_code } from "./clear-code";
+export { onSaveSlackEvent as on_save_slack_event } from "./on-save-slack-event";
 export { slackRedirect as redirect } from "./redirect";
 export { slackReply as reply } from "./reply";
 export { verifySlackCode as verify_code } from "./verify-code";
